+++ public/files/2048/js/application.js
@@ -1,1 +1,0 @@
-animationDelay=100,minSearchTime=100,window.requestAnimationFrame((function(){new GameManager(4,KeyboardInputManager,HTMLActuator)}));