+++ src/components/nav.jsx
@@ -25,13 +25,13 @@ function Nav() {
         )}
       </Link>
       <div className="navitems">
-        <Link className="navitem" to="/apps.html">
+        <Link className="navitem" to="/1.html">
           <AppsIcon />
         </Link>
-        <Link className="navitem" to="/games.html">
+        <Link className="navitem" to="/2.html">
           <VideogameAssetIcon />
         </Link>
-        <Link className="navitem" to="/settings/search.html">
+        <Link className="navitem" to="/3/1.html">
           <SettingsIcon />
         </Link>
       </div>